feat: support for Pure Storage RapidFiles Toolkit v2 (#2419)
Backported-from: main (24.09)
Backported-to: 24.03
Backport-of: 2419
