+++ changes/2419.feature.md
@@ -1,0 +1,1 @@
+Add support for PureStorage RapidFiles Toolkit v2
